add styling for book storage and book display
The book storage container is set to be a grid container that hast three columns. The book display has been styled so when books are added the styling is applied. Set font in "#login-button" id in styles.css to be consistent with the rest of the pages. Commit name "Book Apperance and Storage Commit".