add modal pop up when "Add Book" is clicked.
When the "Add Book" button is clicked then a modal with a menu shows up for the user to add a book. The "script.js" file has also been created. The dialog tag with id "add-book-dialog-container" has been created in index.html for the modal pop up. The styles.css file has been update with styles for the new dialog element.