add "Remove" and "Read"/"Not read" buttons
The remove button on the book now removes the book object from the javascript array which will be removed from the page when the book storage is refreshed and updated. The read/not read button now changes color based on the read status of the book and it updates the status on the corresponding book object.